+++ src/MatrixGenerator.cc
@@ -1,10 +1,10 @@
 #include "MatrixGenerator.h"
 
-std::vector<std::vector<int>> MatrixGenerator::generator(const int rows, const int columns, const int range) {
+std::vector<std::vector<long long>> MatrixGenerator::generator(const int rows, const int columns, const int range) {
   //inicializamos la semilla
   srand(time(nullptr));
 
-  std::vector<std::vector<int>> matriz_generada (rows, std::vector<int>(columns));
+  std::vector<std::vector<long long>> matriz_generada (rows, std::vector<long long>(columns));
 
   for (int i = 0; i < rows; i++) {
     for (int j = 0; j < columns; j++) {
